+++ docs/changelog.rst
@@ -27,6 +27,11 @@ New features
   :conf:`plugins.musicbrainz:aliases_as_credits` to make
   aliases-as-artist-credit optional.
 - :doc:`plugins/badfiles`: Added settings for auto error and warning actions.
+- :doc:`plugins/tidal`: New ``item_types`` flexible attributes
+  ``tidal_track_id``, ``tidal_album_id``, ``tidal_artist_id``,
+  ``tidal_track_popularity``, ``tidal_alb_popularity``, and ``tidal_updated``
+  are now populated during import. Added new ``beet tidalsync`` command to sync
+  popularity data for imported items.
 
 Bug fixes
 ~~~~~~~~~